// src/type.ts
export type Product = {
  id: number;
  title: string;
  price: number;
  category: string;
  description: string;
  image: string;
  rating: {
    rate: number;
  };
};

export type Cart = {
  id: number;
  userId: number;
  date: number;
  products: [];
};

export type User = {
  id: number;
  email: string;
  username: string;
  name: {
    firstname: string;
    lastname: string;
  };
  address: {
    city: string;
    street: string;
  };
};